Reorganised MPI-based Python tests to run directly from pytest.
