+++ testsuite/pytests/mpi/2/connect_test_base.py
@@ -1,0 +1,1 @@
+../../connect_test_base.py
