+++ testsuite/pytests/mpi/2/test_connect_all_to_all.py
@@ -1,0 +1,1 @@
+../../test_connect_all_to_all.py
