+++ testsuite/pytests/mpi/2/test_connect_fixed_outdegree.py
@@ -1,0 +1,1 @@
+../../test_connect_fixed_outdegree.py
