+++ testsuite/pytests/mpi/2/test_connect_one_to_one.py
@@ -1,0 +1,1 @@
+../../test_connect_one_to_one.py
