+++ testsuite/pytests/mpi/2/test_connect_symmetric_pairwise_bernoulli.py
@@ -1,0 +1,1 @@
+../../test_connect_symmetric_pairwise_bernoulli.py
