+++ testsuite/pytests/test_connect_all_to_all.py
@@ -23,12 +23,11 @@
 import unittest
 import numpy as np
 import scipy.stats
-import test_connect_helpers as hf
-from test_connect_parameters import TestParams
+import connect_test_base as hf
 
 
 @hf.nest.ll_api.check_stack
-class TestAllToAll(TestParams):
+class TestAllToAll(hf.ConnectTestBase):
 
     # specify connection pattern
     rule = 'all_to_all'
