+++ testsuite/pytests/test_connect_fixed_total_number.py
@@ -22,11 +22,10 @@
 import numpy as np
 import unittest
 import scipy.stats
-import test_connect_helpers as hf
-from test_connect_parameters import TestParams
+import connect_test_base as hf
 
 
-class TestFixedTotalNumber(TestParams):
+class TestFixedTotalNumber(hf.ConnectTestBase):
 
     # specify connection pattern and specific params
     rule = 'fixed_total_number'
