+++ testsuite/pytests/test_connect_symmetric_pairwise_bernoulli.py
@@ -24,11 +24,10 @@ import collections
 import numpy as np
 import unittest
 import scipy.stats
-import test_connect_helpers as hf
-from test_connect_parameters import TestParams
+import connect_test_base as hf
 
 
-class TestSymmetricPairwiseBernoulli(TestParams):
+class TestSymmetricPairwiseBernoulli(hf.ConnectTestBase):
 
     # sizes of source-, target-population and connection probability for
     # statistical test
